Move getDepthSettingsFromFileUI into engine/src/depth/panelState.ts
Pure DOM-form reader with no instance-state dependency; joins the other
depth panel/form helpers already extracted there.
